Refactor imageUtils to radarImageUtils folder
